requirements.txt
asgiref==3.8.1
Django==5.1
django-phonenumber-field==8.0.0
sqlparse==0.5.1
tzdata==2024.1
